full-trace.sh: set userspace probes on entry and exit
This commit allows to set userspace probes for each function entry and
exit. Function exits are found by disassembling the function and grepping
for a return instruction. This commit also modifies the format of the
symbol tables (files /tmp/p_*). A third column specifies the symbol
as an entry or an exit point.