full-trace.sh: trim program basename in write_trace() to seven characters
The write_trace() function uses the program basename to find the
last occurrence of the program name in a context switch. Unfortunately,
that same name is trimmed to seven characters in the trace output.
As a consequence, grep is unable to find the last occurrence of the
program if its name is longer than seven characters. The following
commit fixes the problem, using only the first seven characters of
the program name in the comparison.