full-trace.sh: really pass command basename to write_trace()
The write_trace() function needs the executed command's basename
to trim the trace. Previously, the function's second argument
was the basename of the whole command line (an operation whose
correctness depended on the command line's format). This patch
correctly extracts the command basename and passes it to the
write_trace() function.