full-trace.sh: trim trace to "PROCESS START" only if marker exists
The write_trace() function uses the sed editor to delete from the
trace the initial part, not pertaining to the execution of the
traced process. However, if the marker was not recorded for some
reason, sed scans the whole trace trimming line after line and
searching for the matching line. This results in the whole trace
being deleted. This patch adds a check to run sed only if the
"PROCESS START" marker exists.