full-trace.sh: first attempt at adding the duration of userspace functions
ftrace is able to add absolute timestamps of uprobe events to the trace
output, but it does not compute the duration of recorded userspace functions,
as it has no notion of entry and exit points. This commit introduces an awk
script that tracks entry and exit points of userspace functions and computes
the duration of their execution.
LIMITATIONS: as of now, function execution times do not work with recursive
functions and with multithreaded programs.