full-trace.sh: use process shortname to remove spurious uprobe events
Previously, the remove-spurious-uprobes() function used the CPU number
to filter out uprobe events that didn't execute on the same CPU as
the main process. While tracing a multithreaded application, this
leads to randomly deleting events from threads belonging to the
traced application. This commit lets the function use, instead, the
shortname of the process (that is the process name truncated to seven
characters max).