full-trace.sh: use split command with "-l" option
The full-trace.sh script uses the split command to divide the original
ftrace output in chunks according to the number of CPUs; this is
implemented by using the split command with its "-n" option, that
accepts as a parameter the number of chunks.
Unfortunately, some versions of the split command don't support such
an option. Use the "-l" option instead, that accepts as a parameter
the number of lines per chunk; dynamically compute the number of lines
to be kept in each chunk according to the total number of lines of the
original ftrace output and the number of CPUs.