fulltrace.sh: changed the buffer's heuristic to comply with ftrace
ftrace uses the number of possible CPUs instead of the number of present CPUs
to compute the size of the per CPU ring buffer. "Possible" refers to the number
of CPUs supported by the chipset, "present" refers instead to the number of
physically present CPUs in the system (which can be both online and offline).
On some machines these values are different, so we need to consider the actual
number of possible CPUs.