full-trace.sh: check if the mandatory argument has been given
The full-trace.sh script needs to know which command is to be
traced. However, no check is performed on the presence of such
an argument. This results in non-explicative errors during the
execution of the script. This patch adds a check on the presence
of this mandatory argument and possibly notifies the user about
its absence, printing also the script usage.