README.md: fix TOC and add a paragraph on ftrace buffer sizing
The README.md file contains a table of contents, whose bullets are
links to the specific paragraphs. The file used anchor tags to define
the internal hyperlinks; however, the special github Markdown plugins
automatically define anchor points for each paragraph. This patch
attempts to adapt the TOC to play along nicely with the auto-generated
anchor points.
This patch also adds a brief paragraph and an usage example for the
-b option of the full-trace.sh script, that allows the user to specify
the per-CPU ftrace buffer.