full-trace.sh: check configuration of running kernel
The fulltrace utility relies on the presence of some features built
in the current kernel. This patch lets the full-trace.sh script
check whether the kernel configuration includes all the needed
options. The script initially tries to retrieve the configuration of
the running kernel from the /proc/config.gz interface. If this file does
not exist, a second attempt is done to find the configuration as the
/boot/config-<running kernel version> file. If none of these attempts
succeeds, fulltrace terminates. If a valid configuration file is found,
fulltrace may either run successfully (if all the needed options are
found to be present) or exit with a list of missing configuration
options.
This patch also adds a "-i" option that allows the user to ignore the
configuration check. It is however not possible to skip the check if
the uprobes subsystem is missing in the debugfs.

Reported-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>