full-trace.sh: fix regexp in remove-spurious-uprobes()
The full-trace.sh script uses a regexp to identify uprobe events
in the ftrace output. The regexp, however, did not correctly match
uprobe events for two reasons: 1) it had extra spaces; 2) it checked
only for C-styled comments, which are also present in some ordinary
ftrace events. This patch fixes the issues by using a better regexp
and removing the extra spaces.

Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>
Reviewed-by: Serena Ziviani <senseriumi@gmail.com>