full-trace.sh: clear set_ftrace_pid tunable in ftrace_off()
The full-trace.sh script does not clear the PID filter of the ftrace
subsystem before exiting. This does not result in any error of the
fulltrace utility, as the filter is always cleared and re-set when
a trace is started, but it may give issues if someone uses the ftrace
subsystem for other purposes after running fulltrace. This patch
lets the ftrace_off() function clear the PID filter.

Reported-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>