full-trace.sh: check if given executable exists
The full-trace.sh script uses the which command to determine the
full path of the executable that must be traced. However, it does
not perform any check on the existence of such an executable.
This may result in an erratic behavior of the script and in non
explicative errors. This patch attempts to fix the issue by checking
whether the which command actually finds the given executable and
printing the usage of the full-trace.sh script instead of an error
message of the which command.

Reported-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>