full-trace.sh: move the prompt for password after the initial checks
This patch attempts to reduce the number of cases in which a password
request is done right before a program failure.
This is done by moving "sudo -v" command from before to after the check
on the existence of the mandatory argument.
This patch also moves a variable declaration that also causes this behaviour.

Signed-off-by: Serena Ziviani <senseriumi@gmail.com>
Reviewed-by: Arianna Avanzini <avanzini.arianna@gmail.com>