full-trace.sh: check if the user wants to exit on "sudo -v" command
This patch adds a control on the exit state of "sudo -v" command
and make full-trace.sh exits if it's abnormal.
Before, an exit attempt on the first password prompt resulted in
a multiple password check due to the continuation of the program.
Now, an exit attempt on the first password propt results in
the termination of the program with code 1.

Signed-off-by: Serena Ziviani <senseriumi@gmail.com>
Reviewed-by: Arianna Avanzini <avanzini.arianna@gmail.com>